Type four call sites to the prototype they had been hiding
Same -W all audit as the previous commit, for the four units where the
missing declaration was also masking an argument-type mismatch, so the
include alone does not compile.

ObjGroup_AddObject takes an int object handle, and both 47.c and
DIMSnowHorn.c already spell the sibling ObjGroup_RemoveObject call
(int)obj -- the Add call now matches. vecRotateZXY takes s16*, and
WCPushBlockRotationWork opens with the yaw/pitch/roll triple it wants.
GXGetScissor writes u32, so fearTestMeterDraw and hudDrawAirMeter declare
their saved scissor rectangles u32 instead of int.

All four objects are byte-identical to their previous build in section
content and relocations; ninja all_source EXIT=0 and main.dol: OK.

+++ src/dlls/engine/47/47.c
@@ -12,6 +12,7 @@
 #include "main/frame_timing.h"
 #include "main/track_dolphin_api.h"
 #include "main/obj_message.h"
+#include "main/obj_group.h"
 
 #define PAD_BUTTON_A              0x100
 #define CARRYABLE_MSG_PLAYER_GRAB 0x100008
@@ -287,7 +288,7 @@ int Carryable_updateHeld(GameObject* obj, void* state)
 void Carryable_init(GameObject* obj, void* state, int arg2)
 {
     CarryableUpdateHeldState* s = (CarryableUpdateHeldState*)state;
-    ObjGroup_AddObject(obj, CARRYABLE_OBJGROUP);
+    ObjGroup_AddObject((int)obj, CARRYABLE_OBJGROUP);
     s->unk2 = 0;
     s->carryState = CARRY_STATE_RESTING;
     s->pad4[0] = 0;
